User with valid JWT token can only access their own tasks

+++ src/main/java/mark/todo/web/UserController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.annotation.security.RolesAllowed;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.Set;
 
 
@@ -34,7 +35,7 @@ public class UserController {
     private JwtTokenUtil jwtTokenUtil;
 
     @GetMapping("/{id}")
-    @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
+    @RolesAllowed({"ROLE_ADMIN"})
     public ResponseEntity<?> findById(@PathVariable Long id) {
         return new ResponseEntity<>(userService.getUser(id), HttpStatus.OK);
     }
@@ -60,7 +61,7 @@ public class UserController {
 
             //generate token and send back
             String accessToken = jwtTokenUtil.generateAccessToken(user);
-            AuthResponse response = new AuthResponse(user.getId(), accessToken);
+            AuthResponse response = new AuthResponse(user.getId(), request.getUsername(), accessToken);
 
             return ResponseEntity.accepted().body(response);
         } catch (BadCredentialsException exception) {
@@ -75,7 +76,10 @@ public class UserController {
     //adding task
     @PostMapping("/{id}/addTask")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
-    public ResponseEntity<?> addTask(@PathVariable Long id, @Valid @RequestBody Task task) {
+    public ResponseEntity<?> addTask(@PathVariable Long id, @Valid @RequestBody Task task, Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userService.createTask(id, task);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -83,21 +87,30 @@ public class UserController {
     //display tasks
     @GetMapping("/{id}/getTasks")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
-    public ResponseEntity<Set<Task>> getTasks(@PathVariable Long id) {
+    public ResponseEntity<Set<Task>> getTasks(@PathVariable Long id, Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(userService.allTasks(id), HttpStatus.OK);
     }
 
     //display completed tasks
     @GetMapping("/{id}/completedTasks")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
-    public ResponseEntity<Set<Task>> getCompletedTasks(@PathVariable Long id) {
+    public ResponseEntity<Set<Task>> getCompletedTasks(@PathVariable Long id, Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(userService.finishedTasks(id), HttpStatus.OK);
     }
 
     //edit task
     @PutMapping("/{id}/{taskNumber}/edit")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
-    public ResponseEntity<?> editTask(@PathVariable Long id, @PathVariable int taskNumber, @Valid @RequestBody Task task) {
+    public ResponseEntity<?> editTask(@PathVariable Long id, @PathVariable int taskNumber, @Valid @RequestBody Task task, Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userService.editTask(id,taskNumber,task);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -105,7 +118,10 @@ public class UserController {
     //set task to complete
     @PutMapping("/{id}/{taskNumber}/completed")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
-    public ResponseEntity<?> finishTask(@PathVariable Long id, @PathVariable int taskNumber) {
+    public ResponseEntity<?> finishTask(@PathVariable Long id, @PathVariable int taskNumber,Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userService.finishTask(id,taskNumber,true);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -114,7 +130,10 @@ public class UserController {
     @DeleteMapping("/{id}/{taskNumber}/delete")
     @RolesAllowed({"ROLE_ADMIN","ROLE_USER"})
     @Transactional
-    public ResponseEntity<?> deleteTask(@PathVariable Long id, @PathVariable int taskNumber) {
+    public ResponseEntity<?> deleteTask(@PathVariable Long id, @PathVariable int taskNumber,Principal p) {
+        if(!p.getName().toString().equals(userService.getUser(id).getUsername())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userService.deleteTask(id, taskNumber);
         return new ResponseEntity<>(HttpStatus.OK);
     }
